Fix dev container compatibility in favor of podman
Podman throws an error when using the `--device-cgroup-rule` option in a
rootless container. However, docker requires this rule to allow using USB
devices such as the ST-Links. I've tried every method I can think of to
make a dev container config which works with both docker and podman,
while retaining the ability to flash, but this doesn't seem doable.

With that in mind, I have setup the dev container in favor of podman,
while breaking compatibility with docker (due to `--userns=keep-id`).
This simplifies installation greatly, with the stipulation that users install
podman in order to use this container.